# openapi.yaml
openapi: 3.0.0
info:
  title: HandyNear API (Team 7)
  version: 1.0.0
  description: Handyman service discovery and booking management API for HandyNear.
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
  /api/handymen:
    get:
      summary: List all available handymen
      parameters:
        - name: serviceCategory
          in: query
          required: false
          schema:
            type: string
            example: "Plumbing"
          description: Filter providers by specific trade or skill category.
        - name: available
          in: query
          required: false
          schema:
            type: boolean
            example: true
          description: Filter only handymen currently accepting jobs.
      responses:
        '200':
          description: A list of available handymen matching criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Handyman'
        '400':
          description: Invalid query parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/handymen/{id}:
    get:
      summary: Get details for a specific handyman
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 12
          description: Unique identifier of the handyman.
      responses:
        '200':
          description: Detailed profile of the specified handyman.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Handyman'
        '404':
          description: Handyman not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/bookings:
    get:
      summary: Retrieve all bookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
            example: "confirmed"
          description: Filter bookings by lifecycle status.
      responses:
        '200':
          description: List of bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
    post:
      summary: Create a new service booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Validation error or malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/bookings/{id}:
    get:
      summary: Retrieve a single booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 101
          description: Unique identifier of the booking record.
      responses:
        '200':
          description: Booking details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update booking status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 101
          description: Unique identifier of the booking to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, completed, cancelled]
                  example: "completed"
                notes:
                  type: string
                  example: "Work completed ahead of schedule."
      responses:
        '200':
          description: Booking successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking record not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Handyman:
      type: object
      required:
        - id
        - name
        - serviceCategory
        - hourlyRate
        - isAvailable
        
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: "John Kamau"
        phone:
          type: string
          example: "+254712345678"
        email:
          type : string
          format: email
          example: "johnkamau@example.com"
        nationalID:
          type: string
          pattern: '^[0-9]{8,20}$'
          exapmle: "3458933254"
        serviceCategory:
          type: string
          example: "Plumbing"
        hourlyRate:
          type: number
          format: float
          example: 1500.00
        rating:
          type: number
          format: float
          example: 4.8
        isAvailable:
          type: boolean
          example: true
        location:
          type: string
          example: "Nairobi"

    CreateBookingRequest:
      type: object
      required:
        - handymanId
        - customerName
        - customerPhone
        - serviceDescription
        - scheduledTime
      properties:
        handymanId:
          type: integer
          example: 12
        customerName:
          type: string
          example: "Sarah Mwangi"
        customerPhone:
          type: string
          example: "+254798765432"
        serviceDescription:
          type: string
          example: "Leaking kitchen sink faucet repair"
        scheduledTime:
          type: string
          format: date-time
          example: "2026-09-12T14:30:00Z"

    Booking:
      type: object
      required:
        - id
        - handymanId
        - customerName
        - customerPhone
        - serviceDescription
        - scheduledTime
        - status
        - createdAt
      properties:
        id:
          type: integer
          example: 101
        handymanId:
          type: integer
          example: 12
        customerName:
          type: string
          example: "Sarah Mwangi"
        customerPhone:
          type: string
          example: "+254798765432"
        customerEmail:
          type: string
          format : email
          example: sarah@example.com
        serviceDescription:
          type: string
          example: "Leaking kitchen sink faucet repair"
        scheduledTime:
          type: string
          format: date-time
          example: "2026-09-12T14:30:00Z"
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
          example: "confirmed"
        notes:
          type: string
          example: "Customer requested arrival after 2:30 PM."
        createdAt:
          type: string
          format: date-time
          example: "2026-09-09T09:15:00Z"

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "RESOURCE_NOT_FOUND"
        message:
          type: string
          example: "The requested resource could not be found."